requirements.txt
django>=4.2.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
torch>=2.0.0
open-clip-torch>=2.20.0
albumentations>=1.3.0
joblib>=1.3.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
xlsxwriter>=3.1.0